+++ pill/lib/model/search/dur_search_result.dart
